refactor: change menus structure
