+++ Hospital/src/main/java/Main.java
@@ -1,18 +1,7 @@
 package main.java;
 
-import main.java.components.Appointment;
-import main.java.controllers.menus.DoctorMenu;
-import main.java.controllers.menus.NurseMenu;
-import main.java.controllers.menus.PatientMenu;
-import main.java.controllers.resource_controllers.DataBaseIO;
-import main.java.users.Patient;
-import main.java.users.leadership.ChiefPhysician;
-import main.java.users.stuff.Doctor;
-import main.java.users.stuff.Nurse;
-import main.java.usersdb.DoctorDB;
+import main.java.controllers.menus.user_menus.NurseMenu;
 import main.java.usersdb.NurseDB;
-import main.java.usersdb.PatientDB;
-import main.java.controllers.resource_controllers.Authorizer;
 
 public class Main {
     public static void main(String[] args) {
